refactor : event/index.ts 파일명 event/movieEventHandler.ts 으로 변경
